fix(preview-sidebar): move code to SidebarPanels
