+++ src/elements/content-sidebar/SidebarPanels.js
@@ -6,7 +6,7 @@
 
 import * as React from 'react';
 import flow from 'lodash/flow';
-import { Redirect, Route, Switch } from 'react-router-dom';
+import { matchPath, Redirect, Route, Switch, type Location } from 'react-router-dom';
 import SidebarUtils from './SidebarUtils';
 import withSidebarAnnotations from './withSidebarAnnotations';
 import { withAnnotatorContext } from '../common/annotator-context';
@@ -18,7 +18,6 @@ import {
     ORIGIN_METADATA_SIDEBAR,
     ORIGIN_SKILLS_SIDEBAR,
     ORIGIN_VERSIONS_SIDEBAR,
-    SIDEBAR_PATH_VERSIONS,
     SIDEBAR_VIEW_ACTIVITY,
     SIDEBAR_VIEW_DETAILS,
     SIDEBAR_VIEW_METADATA,
@@ -48,6 +47,7 @@ type Props = {
     hasSkills: boolean,
     hasVersions: boolean,
     isOpen: boolean,
+    location: Location,
     metadataSidebarProps: MetadataSidebarProps,
     onAnnotationSelect?: Function,
     onVersionChange?: Function,
@@ -88,6 +88,8 @@ const LoadableVersionsSidebar = SidebarUtils.getAsyncSidebarContent(
     MARK_NAME_JS_LOADING_VERSIONS,
 );
 
+const SIDEBAR_PATH_VERSIONS = '/:sidebar(activity|details)/versions/:versionId?';
+
 class SidebarPanels extends React.Component<Props, State> {
     activitySidebar: ElementRefType = React.createRef();
 
@@ -102,6 +104,21 @@ class SidebarPanels extends React.Component<Props, State> {
     componentDidMount() {
         this.setState({ isInitialized: true });
     }
+
+    componentDidUpdate(prevProps: Props): void {
+        const { location, onVersionChange }: Props = this.props;
+        const { location: prevLocation }: Props = prevProps;
+
+        // Reset the current version id if the wrapping versions route is no longer active
+        if (onVersionChange && this.getVersionsMatchPath(prevLocation) && !this.getVersionsMatchPath(location)) {
+            onVersionChange(null);
+        }
+    }
+
+    getVersionsMatchPath = (location: Location) => {
+        const { pathname } = location;
+        return matchPath(pathname, SIDEBAR_PATH_VERSIONS);
+    };
 
     /**
      * Refreshes the contents of the active sidebar
